Don't use jetpack unless you are high in the air.
This prevents one class of accidental jetpack use, when you are about to
land and want to jump again.  It also prevents the case where you walk
off a corner and jump, but only if the ground is nearby.  If you walk
off a cliff and then try to jump it will still use the jetpack.