Rudimentary fix for wrapping.
Force the player to stay within the nominal x range (0 - pixel_width).
This mostly resolves the issues with wrapping.  There are still some
edge cases where the player is near the "seam" of the world.
Additionally, the starfield background doesn't loop at all so it
suddenly changes whenever the player crosses the seam.